+++ THRCoreData/ManagedObjectObserver.swift
@@ -9,22 +9,25 @@
 import CoreData
 
 public enum ManagedObjectChangeType {
+    case initialised
     case refreshed
     case updated
     case deleted
 }
 
 /// Observe changes made to a managed object (refreshed, updated, deleted).
-open class ManagedObjectChangeObserver<T> where T: NSManagedObject & ManagedObjectType {
+open class ManagedObjectObserver<T>: NSObject where T: NSManagedObject & ManagedObjectType {
     
     public typealias OnChange = ((T, ManagedObjectChangeType) -> Void)
     
-    public var onChange: OnChange?
     public var enabled: Bool = true
     public let object: T
 
     private let context: NSManagedObjectContext
     private let managedObjectID: NSManagedObjectID
+    
+    private var notifierRunning: Bool = false
+    private var onChange: OnChange?
 
     /// Create a new ManagedObjectChangeObserver.
     /// The object will be observed in its original managedObjectContext.
@@ -53,16 +56,26 @@ open class ManagedObjectChangeObserver<T> where T: NSManagedObject & ManagedObje
         self.managedObjectID = managedObjectID
         self.context = context
         self.object = context.object(with: managedObjectID) as! T
+        super.init()
+    }
+    
+    public func startObserving(_ onChange: OnChange?) {
+        guard !notifierRunning else { return }
+
+        self.onChange = onChange
         
         NotificationCenter.default.addObserver(forName: NSNotification.Name.NSManagedObjectContextObjectsDidChange, object: nil, queue: nil) { [weak self] (note) in
             guard let strongSelf = self else { return }
             guard strongSelf.enabled else { return }
             let notification = ObjectsDidChangeNotification(notification: note)
-            guard notification.managedObjectContext == context else { return }
+            guard notification.managedObjectContext == strongSelf.context else { return }
             strongSelf.checkForMatchingObject(in: notification.refreshedObjects, changeType: .refreshed)
             strongSelf.checkForMatchingObject(in: notification.updatedObjects, changeType: .updated)
             strongSelf.checkForMatchingObject(in: notification.deletedObjects, changeType: .deleted)
         }
+        
+        onChange?(object, .initialised)
+        notifierRunning = true
     }
     
     private func checkForMatchingObject(in changedObjects: Set<NSManagedObject>, changeType: ManagedObjectChangeType) {
@@ -77,9 +90,9 @@ extension ManagedObjectType where Self: NSManagedObject {
     ///
     /// - Parameter onChange: A callback called when the object is changed.
     /// - Returns: A ManagedObjectChangeObserver initialised with self as the managed object.
-    public func observe(onChange: @escaping ManagedObjectChangeObserver<Self>.OnChange) -> ManagedObjectChangeObserver<Self> {
-        let observer = ManagedObjectChangeObserver<Self>(managedObject: self)
-        observer.onChange = onChange
+    public func observe(onChange: @escaping ManagedObjectObserver<Self>.OnChange) -> ManagedObjectObserver<Self> {
+        let observer = ManagedObjectObserver<Self>(managedObject: self)
+        observer.startObserving(onChange)
         return observer
     }
     
@@ -89,9 +102,9 @@ extension ManagedObjectType where Self: NSManagedObject {
     ///   - context: The context that will hold the fetched object. It will first be fetched from this context, so it may differ from its original.
     ///   - onChange: A callback called when the object is changed.
     /// - Returns: A ManagedObjectChangeObserver initialised with self as the managed object.
-    public func observe(in context: NSManagedObjectContext, onChange: @escaping ManagedObjectChangeObserver<Self>.OnChange) -> ManagedObjectChangeObserver<Self> {
-        let observer = ManagedObjectChangeObserver<Self>(managedObject: self, context: context)
-        observer.onChange = onChange
+    public func observe(in context: NSManagedObjectContext, onChange: @escaping ManagedObjectObserver<Self>.OnChange) -> ManagedObjectObserver<Self> {
+        let observer = ManagedObjectObserver<Self>(managedObject: self, context: context)
+        observer.startObserving(onChange)
         return observer
     }
 }
@@ -104,9 +117,9 @@ extension NSManagedObjectID {
     ///   - context: The context that will hold the fetched object.
     ///   - onChange: A callback called when the object is changed.
     /// - Returns: A ManagedObjectChangeObserver initialised with the managed object referred to by the ID.
-    public func observe<T>(in context: NSManagedObjectContext, onChange: @escaping ManagedObjectChangeObserver<T>.OnChange) -> ManagedObjectChangeObserver<T> {
-        let observer = ManagedObjectChangeObserver<T>(managedObjectID: self, context: context)
-        observer.onChange = onChange
+    public func observe<T>(in context: NSManagedObjectContext, onChange: @escaping ManagedObjectObserver<T>.OnChange) -> ManagedObjectObserver<T> {
+        let observer = ManagedObjectObserver<T>(managedObjectID: self, context: context)
+        observer.startObserving(onChange)
         return observer
     }
 }